# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: UptivaLab API
  description: |
    Comprehensive uptime monitoring and status page platform API.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Features
    - HTTP/HTTPS monitoring
    - TCP port monitoring
    - Ping monitoring
    - DNS monitoring
    - Certificate expiry tracking
    - Synthetic browser testing
    - Docker container monitoring
    - gRPC service monitoring
    - Push-based heartbeat monitoring
    - Incident management
    - Maintenance windows
    - Public status pages
    - Multi-channel notifications
    
  version: 1.0.0
  contact:
    name: UptivaLab Support
    url: https://github.com/tekuonline/uptivaLab
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.uptivalab.com/api
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /status/public/{slug}:
    get:
      tags: [Status]
      summary: Get public status page
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public status page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicStatus'

  # Heartbeats
...
  /status-pages:
    get:
      tags: [Status Pages]
      summary: List status pages
      responses:
        '200':
          description: List of status pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatusPage'

    post:
      tags: [Status Pages]
      summary: Create status page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusPageInput'
      responses:
        '200':
          description: Status page created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusPage'
  /status-pages/{id}:
    get:
      tags: [Status Pages]
      summary: Get status page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Status page details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusPage'

    put:
      tags: [Status Pages]
      summary: Update status page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusPageInput'
      responses:
        '200':
          description: Status page updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusPage'

    delete:
      tags: [Status Pages]
      summary: Delete status page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Status page deleted
  /public/status/{slug}:
    get:
      tags: [Status Pages]
      summary: Get public status page
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public status page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicStatusPage'

  # Settings
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Monitor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [http, tcp, ping, dns, certificate, docker, grpc, synthetic]
        url:
          type: string
        interval:
          type: integer
          description: Check interval in seconds
        timeout:
          type: integer
          description: Timeout in milliseconds
        config:
          type: object
          additionalProperties: true
        paused:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MonitorWithStatus:
      allOf:
        - $ref: '#/components/schemas/Monitor'
        - type: object
          properties:
            latestCheck:
              $ref: '#/components/schemas/CheckResult'
            meta:
              type: object
              properties:
                certificateExpiresAt:
                  type: string
                  format: date-time
                certificateDaysLeft:
                  type: integer
    CheckResult:
      type: object
      properties:
        id:
          type: string
        monitorId:
          type: string
        status:
          type: string
          enum: [up, down]
        message:
          type: string
        latencyMs:
          type: integer
        checkedAt:
          type: string
          format: date-time
        payload:
          type: object
          additionalProperties: true
...
    StatusSnapshot:
      type: object
      properties:
        monitor:
          $ref: '#/components/schemas/Monitor'
        status:
          type: string
          enum: [up, down]
        lastCheck:
          type: string
          format: date-time
        uptime:
          type: number
          format: float
...
    Incident:
      type: object
      properties:
        id:
          type: string
        monitorId:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [investigating, identified, monitoring, resolved]
        severity:
          type: string
          enum: [minor, major, critical]
        startedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    MaintenanceWindow:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        monitorIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    StatusPage:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        monitorIds:
          type: array
          items:
            type: string
        public:
          type: boolean
        createdAt:
          type: string
          format: date-time
    StatusPageInput:
      type: object
      required: [title, slug, monitorIds]
      properties:
        title:
          type: string
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        description:
          type: string
        monitorIds:
          type: array
          items:
            type: string
        public:
          type: boolean
          default: true
    PublicStatusPage:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        monitors:
          type: array
          items:
            $ref: '#/components/schemas/MonitorWithStatus'
        incidents:
          type: array
          items:
            $ref: '#/components/schemas/Incident'
        maintenance:
          type: array
          items:
            $ref: '#/components/schemas/MaintenanceWindow'
    PublicStatus:
      type: object
      properties:
        overall:
          type: string
          enum: [operational, degraded, down]
        monitors:
          type: array
          items:
            $ref: '#/components/schemas/StatusSnapshot'
...